+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-## 2.20.0 (Unreleased)
+## 2.20.1 (Unreleased)
+BUG FIXES:
+* Added validations to fail if no trigger conditions defined in monitor resource (GH-463)
+
 FEATURES:
-* Add new optional `min_data_points` field to metrics monitor trigger_conditions
+* Add new optional `min_data_points` field to metrics monitor trigger_conditions (GH-464)
+
+## 2.20.0 (December 15, 2022)
+FEATURES:
+* **New Resource:** sumologic_cse_custom_match_list_column (GH-462)
 
 BUG FIXES:
 * Fix typo on cse_match_list documentation (GH-461)